+++ packages/api-proxy/src/v3/polygon/feewallet/fee-wallet.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from "@nestjs/common";
+import { SDK } from "@haechi-labs/henesis-wallet-core";
+import { BalanceDTO } from "../../eth/dto/balance.dto";
+
+@Injectable()
+export class FeeWalletService {
+  public async getBalance(sdk: SDK): Promise<BalanceDTO> {
+    return BalanceDTO.fromBalance(
+      await sdk.polygon.henesisKeys.getHenesisKeyBalance()
+    );
+  }
+}